MGMT-4179: Move KubeAPI types to under the top-level directory
This change migrates kubeAPI types to be a top-level directory.
Imports and docs were updated to match the new location.
