+++ internal/controller/controllers/agentserviceconfig_controller.go
@@ -27,8 +27,8 @@ import (
 	"github.com/go-openapi/swag"
 	"github.com/hashicorp/go-version"
 	routev1 "github.com/openshift/api/route/v1"
+	aiv1beta1 "github.com/openshift/assisted-service/api/v1beta1"
 	"github.com/openshift/assisted-service/internal/common"
-	aiv1beta1 "github.com/openshift/assisted-service/internal/controller/api/v1beta1"
 	"github.com/openshift/assisted-service/internal/gencrypto"
 	"github.com/openshift/assisted-service/models"
 	logutil "github.com/openshift/assisted-service/pkg/log"
